+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Displaying Pokemon Types and IDs
 - Pokemon Fun Facts
+- Filtering pokemons by type in search
 
 ##### Modified